+++ packages/core-js/internals/function-to-string.js
@@ -1,3 +1,0 @@
-var shared = require('../internals/shared');
-
-module.exports = shared('native-function-to-string', Function.toString);