+++ packages/core-js/internals/inspect-source.js
@@ -1,0 +1,7 @@
+var shared = require('../internals/shared');
+
+var functionToString = Function.toString;
+
+module.exports = shared('inspectSource', function (it) {
+  return functionToString.call(it);
+});